kernel: Various 64-bit fixes in memory/process/thread

+++ src/core/hle/kernel/memory.h
@@ -13,9 +13,9 @@ namespace Kernel {
 class VMManager;
 
 struct MemoryRegionInfo {
-    u32 base; // Not an address, but offset from start of FCRAM
-    u32 size;
-    u32 used;
+    u64 base; // Not an address, but offset from start of FCRAM
+    u64 size;
+    u64 used;
 
     std::shared_ptr<std::vector<u8>> linear_heap_memory;
 };
